BIG CHANGES!
For the most part these changes are to the orginization of classes and which
jars and what classloaders they load into.  The main reason for this was so that
the OM implmenetation classes could access all of jetspeed's internal services
for things like persistence and security.

All implementation classes of the OM have been moved into portal so as to avoid 
cross context issues and now the implmentation classes can take full
advantage of all of Jetspeeds services.

The Container source tree has been completely removed, all remaining
classes have been moved into commons along side the OM interfaces
(which is loaded from shared/lib).  This removes the requirment for a portlet 
application needing to include the container.jar in it's WEB-INF/lib.  However, 
the JetpseedContainerServlet still needs to be defined in each portlet application's web.xml.

Import/mapping changes relative to new package names.


git-svn-id: http://svn.apache.org/repos/asf/portals/jetspeed-2/trunk@185262 13f79535-47bb-0310-9956-ffa450edef68