The beginnings of the contnet server.  The servlet is basically nothing right now.
However, the automated deployment system is 85% complete and is generic enough to 
be used with other types of deployment besides themes content.  For example
it will be very easy to write a DeploymentEventListener that wraps the PAM
for automating portlet application deployment.

The javadocs also need to be fleshed out.


git-svn-id: http://svn.apache.org/repos/asf/portals/jetspeed-2/trunk@185439 13f79535-47bb-0310-9956-ffa450edef68