Removed isUserInRole(roleName) because it is now provided by Pluto itself (in its PortletRequestImpl, see issue: PLUTO-46) and is conflicting with our own implementation.
Furthermore, our implementation did a roleName/roleLink lookup from portlet.xml definition (to be able to handle
the fact Pluto initiall didn't do it yet) which was *not* according to the specs.
The specs says HttpServletRequest.isUserInRole(roleName) "must provide the functionality defined by the Servlet Specification 2.3" (PLT.16.3.3, cxxxvi).
This means that roleName must *only* be checked against web.xml definition, which should be done by the ServletContainer itself (hence the removal of this method).

A less nice consequence of this is that potentially different roleNames must be used when mixing Portlet and Servlet functionality within a PortletApplication:
the abstraction of portlet.xml:roleName->roleLink->web.xml:roleName is not available from the HttpServletRequest.isUserInRole(roleName).
Our  Role Security Test portlet did indeed *misuse* this feature by checking the same roleName against portlet and servlet request.
I also fixed that and the portlet now again correctly shows the roles of the logged on user.


git-svn-id: http://svn.apache.org/repos/asf/portals/jetspeed-2/trunk@187761 13f79535-47bb-0310-9956-ffa450edef68