Refactored SSO feature: --> resolved JS2-173, JS2-388
Creating cookies in the LoginModule for any SSO is feasible but cookies can't be directly injected (in pipeline) into an ongoing request.
Instead I introduced the SSOProxyPortlet that does the same functionality. The SSOProxyPortlet authenticates any SSO for the given principal before it redirects the request to the final destination. If the page has any links to SSO protected sites the user won't be prompted since it has already been authenticated.

Another advantage of this approach is that the SSO component is not bound to the portal infrastructure (LoginModule / pipeline) and therefore allows easier integration into other SSO frameworks like JOSSO.

git-svn-id: http://svn.apache.org/repos/asf/portals/jetspeed-2/trunk@354497 13f79535-47bb-0310-9956-ffa450edef68