refactored Page Manager common code to compensate for no multiple inheritance on the database page manager. In retrospect, I think we should consider injecting the InitablePersistenceBrokerDaoSupport and instead of extending.
The problem was in the DatabasePageManager, it appears that when 
Im in the DelegatingPageManager code, I can't call any *real* implementations 
from DatabasePageManager such as getPage or getFolder. It calls into the dummy Delegating ones. 
This is very f'ed up. 

The solution, with code reuse in mind, is PageManagerUtils to share common code
between the two impls. Still not entirely happy with this solution.


git-svn-id: http://svn.apache.org/repos/asf/portals/jetspeed-2/trunk@371338 13f79535-47bb-0310-9956-ffa450edef68