- correct name of jetspeed-component artifact - first pass at Maven2 portal archetype to duplicate Maven1 plugin j2:portal-genapp goal
Here is an example archetype:create invocation to create a custom portal instance:

mvn archetype:create -DarchetypeGroupId=org.apache.portals.jetspeed-2 -DarchetypeArtifactId=portal-archetype -DarchetypeVersion=2.1-dev -DgroupId=myportalgroup -DartifactId=myportal -Dversion=1.0

After running this command, a new directory named 'myportal' will be created. As
with J2, there is a sample settings.xml.sample file that needs to be customized
and placed in ~/.m2 to continue. The expanded archetype uses a reduced set of
settings as used by J2, so a new settings.xml file does not need to be created
for the archetype. The same J2 build commands are supported for the most part,
(the 'full' and 'min' profiles do not apply since the archetype is generated as
a 'min' deployment; the creation of a test database is not supported in the
'test' profile):

> mvn                 # simple build
> mvn -P test         # simple build with tests
> mvn -P tomcat       # build and deploy to tomcat
> mvn -P tomcat,nodb  # build and deploy to tomcat w/o a db update
> mvn -P tomcat,hot   # build followed by hot deploy

I have not implemented the companion component and application archetypes yet.

More documentation to follow separately.



git-svn-id: http://svn.apache.org/repos/asf/portals/jetspeed-2/trunk@387361 13f79535-47bb-0310-9956-ffa450edef68