Automatically clear cached but no longer valid PortletWindow/PortletEntity instances.
A PortletEntity normally holds a reference to a PortletDefinition and its PortletApplication.
When a Portlet Application is restarted/redeployed with changes in its portlet.xml, the registry for it is updated.
Then, cached PortletWindow/PortletEntity instances still holding out-of-sync references to the "old" registry instances
may cause  rendering problems which can only be solved by restarting the portal!

Now, each PortletWindow and its PortletEntity instance will first be checked when retrieved from the cache and removed from it when out-of-sync.

Note: this is a light-weight validation checking if the referenced PortletApplication is registered in the PortletFactory (meaning its active/available).
These checks will have no real performance penalty for "valid" PortletWindows, only for "offline" Portlet Applications they will (need to) be recreated each and every time. 

Warning: this change requires a mandatory change to the Spring assembly definition for the PortletWindowAccessorImpl (in jetspeed-spring.xml).   

git-svn-id: http://svn.apache.org/repos/asf/portals/jetspeed-2/trunk@408514 13f79535-47bb-0310-9956-ffa450edef68