contribution from Dave Gurney
Problem
------------
The Ajax movement API allows portlets to be added to the page.  Some applications may find that having more than one instance of a portlet on the page is inappropriate.  Currently the code within Jetspeed allows "n" instances of the same portlet without any problems.  (I think it's safe to say that this is a valuable "feature")  In our application, duplicate portlets don't make sense and need to be disallowed.

Solution
--------------
The solution is straight forward and involves adding code to the AddPortletAction.java object to prevent a portlet from being added to the page if it is already present.

jetspeed/components/portal/src/java/org/apache/jetspeed/layout/impl/AddPortletAction.java

Changes to existing code
--------------------------------------
The following changes were made to the existing code to accommodate this new capability:

* Add a new flag to control whether or not duplicate portlets are allowed
* Add a new constructor which takes an additional boolean parameter to specify if duplicate portlets should be allowed. (This allows for full backwards compatibility - the default is to allow duplicates.)

* Check the new duplicate flag within the run() method.  If the flag is true, invoke a new method
* Introduced a new method "checkForDuplicatePortlets" that checks to see if the requested portlet is already on the page

* Introduce a new method that will recursively process the fragment tree looking for duplicates


git-svn-id: http://svn.apache.org/repos/asf/portals/jetspeed-2/trunk@415917 13f79535-47bb-0310-9956-ffa450edef68