Patch for JS2-491: Enhance J2 LDAP Security + documentation See: http://issues.apache.org/jira/browse/JS2-491 Patch provided by Davy De Waele who has put a lot of effort into this, with a great enhancement as well as extended documentation as result. Thanks!
NB: 
I've only been able to successfully run the testcases against OpenLDAP. Our "default" ApacheDS fails miserably because of two critical bugs (which I have reported).
So, for the time being (we have to wait for release 1.0.1), ApacheDS cannot reliably be used with Jetspeed.
OpenLDAP is fine though, and Davy reported success with Domino and SunDS as well.

I'll try to update the documentation online asap.

git-svn-id: http://svn.apache.org/repos/asf/portals/jetspeed-2/trunk@483771 13f79535-47bb-0310-9956-ffa450edef68