https://issues.apache.org/jira/browse/JS2-621 Improve SSO Portlet implementation This is an ongoing work, this first patch applies to:
Fix bug in PersistenceBrokerSSOProvider.java with unscrambling passwords
Specifically doing a “.toString()” on a byte[] object. 
As part of this correction we refactored both the scramble and unscramble methods to be a little 
more human readable and debugger friendly. 
 
The second change centers around the WebContentPortlet.doHttpWebContent() method. 
We changed the return type from Boolean to a byte[]. 
This byte[] represents the refactored page that should be sent to the browser. 
It seems that Jetspeed was properly logging in to the remote site, it would receive the 
html back from that site and refactor it in preparation to send back to the browser but 
it would not actually send it back to the browser. It would only return a boolean true 
that indicated that it had successfully received that new data. It never actually sent it to the browser.

We tested these changes with the NY Times example in Jetspeed and seems to work
  
contribution from Joe Kueser and Tom Campbell of Gestalt, LLC


git-svn-id: http://svn.apache.org/repos/asf/portals/jetspeed-2/trunk@484715 13f79535-47bb-0310-9956-ffa450edef68