Problem:
Only one instance of Derby is allowed to be loaded per instance of the Java Virtual Machine (JVM) 
and only a single Derby instance is allowed to access any given database at a time. 
When the Derby instance connects to a database, a lock is created that helps to enforce these rules. 
The only way to remove that lock is to explicitly tear down Derby’s connection to the database. 
While SQL Ant Task is fully capable of launching the Derby embedded database and executing SQL statements against it, 
it is not capable of handling the appropriate Derby shutdown tasks that are required to use Derby properly.

Hence the Derby database used in an ant <SQL> task remains looked until that process terminates – 
any other connection request – whether from a forked or embedded JAVA process – is rejected.

Solution:

While not the best solution from a performance perspective, the easiest and most reliable is to establish the SQL task from 
within a separate (JAVA) process, which -when forked- will upon completion terminate the database and bring it down 
correctly (i.e. remove the lock).

contribution from Hajo Birthelmer

git-svn-id: http://svn.apache.org/repos/asf/portals/jetspeed-2/trunk@499144 13f79535-47bb-0310-9956-ffa450edef68