1. Fix bug with Tomcat not shutting down properly 2. Add safe checks for null in SecurityHelper because of bug logged by Andrew Hall
For the Tomcat shutdown, there are two issues:
- One we see a message in the console 
"INFO: Failed shutdown of Apache Portable Runtime"
- The second is that Tomcat in fact doesn't shutdown completely. Either the
console remains open and/or the next attempt to start Tomcat fails because
the port is still in use.
In any case the Java process associated with Tomcat is still running.

The first issue of the message is really just a poorly worded diagnostic
message from Tomcat, apparently only in 5.5.12 (see also
http://marc.theaimsgroup.com/?l=tomcat-user&m=113307219707673&w=2) .

The second was a bug in the WorkerMontiorImpl.java not stopping the worker
thread correctly. 
Attached is the fix for that.
Note that rather using the deprecated Thread.stop() it closes the thread
programmatically.

patch 2 from Hajo Birthelmer


git-svn-id: http://svn.apache.org/repos/asf/portals/jetspeed-2/trunk@501578 13f79535-47bb-0310-9956-ffa450edef68