Rolling back previous fix for JS2-462: Need to clean up Velocity logs, too many warnings and some errors even   ERROR velocity - VM #includeNestedLinksWithIconNavigation: error : too few arguments to macro. Wanted 2 got 0
Turns out there are many (potential) problems with recursive Velocity macros...
Velocity does have a (still undocumented) #local directive which is supposed to solve those problems.
But... with the current version (1.4) there are still problems with it although they should be solved in the next version (1.5)
See: http://issues.apache.org/jira/browse/VELOCITY-24

For the time being, better be save than sorry so we will have to keep up with the above error message a while longer.


git-svn-id: http://svn.apache.org/repos/asf/portals/jetspeed-2/trunk@509857 13f79535-47bb-0310-9956-ffa450edef68