https://issues.apache.org/jira/browse/JS2-655
Entity Editor has been broken for a long time.
Also, the entity editor is unsecured.
Propose fixing this bug by retrofitting onto a "ajax-direct" pipeline keyed of the /ajax pipeline mapping
Also assign a security behavior to the ajax valve to give it RBAC security, locking out all AJAX calls not authorized by a list of trusted roles 

(Could have sworn there was a jira issue on this one, but i could not find it, sorry if i have created a dupe)

git-svn-id: http://svn.apache.org/repos/asf/portals/jetspeed-2/trunk@510438 13f79535-47bb-0310-9956-ffa450edef68