JIRA issue JS2-659: Compliance with http://www.apache.org/legal/src-headers.html - bump version to 2.1.1-RC - bump bridges dependencies to 1.0.2-RC (AFAIK now also in compliance with http://www.apache.org/legal/src-headers.html)
AFAIK, this commit brings Jetspeed-2 fully inline with the requirements of http://www.apache.org/legal/src-headers.html,
so I'll close J2-659 hereafter.

I'll also tag this "release candidate" 2.1.1-RC, but (as the -RC already indicates) this will not lead to a new release (yet).
There are still *zero* code changes since the 2.1 release (well, except a minor fix for a testcase misbehavior).

If a real 2.1.1 release is deemed required because of important bugs, we can always create a separate 2.1.1-dev branch based on the 2.1.1-RC tag.
But the main larger development will continue on the trunk for which I will bump the version to 2.2-dev shortly.

git-svn-id: http://svn.apache.org/repos/asf/portals/jetspeed-2/trunk@517075 13f79535-47bb-0310-9956-ffa450edef68