Fix for JS2-484: UserDetailsPortlet doesn't "see" newly added roles until after logging out and in again See: http://issues.apache.org/jira/browse/JS2-484
I've added messaging to the RoleDetails and GroupDetails portlets to notify the UserDetailsPortlet to refresh its Roles (or Groups) cache.

Note: for these messages to be picked up, I had to disable the content cache for the UserDetailsPortlet.
We need a way to clear portlet content caches for selected portlets which are not on the current page...

git-svn-id: http://svn.apache.org/repos/asf/portals/jetspeed-2/trunk@545187 13f79535-47bb-0310-9956-ffa450edef68