Fix for: JS2-735 See: https://issues.apache.org/jira/browse/JS2-735
Note: this a to resolve a Jetty 6 issue (probably a bug) which cannot handle session.invalidate() called from a HttpSessionBindingListener.unboundValue(...).

I'm removing the extra, extra safeguard forcing invalidation when the PortletApplicationSessionMonitor is removed from the session itself.
This isn't absolutely needed, but it now allows an application session to "detach" itself from the Portal Session synchronization by removing the Monitor itself (which would be a weird action anyway). 

git-svn-id: http://svn.apache.org/repos/asf/portals/jetspeed-2/trunk@548358 13f79535-47bb-0310-9956-ffa450edef68