[JS2-634] edit_defaults custom portlet mode
Fixed a bug of auto-dispatching option. The problem was that a portlet can show edit_defaults icon when it does not even EDIT mode.
DecoratorValve now checks if a portlet supports edit mode in auto-dispatching option.
Also, two edit pages for bookmark portlets are updated because the edit pages can be used for both edit and edit_defaults mode now.

git-svn-id: http://svn.apache.org/repos/asf/portals/jetspeed-2/trunk@581786 13f79535-47bb-0310-9956-ffa450edef68