[JS2-764] There are much DOCTYPE directives in every result page of portal
Thorsten Berger pointed out the following problem and suggested solutions:

Unfortunately the DOCTYPE in the decorator's header.vm is not recognized by Internet Explorer, as it's not the first content in the html page: There's one comment (originating from the layout.vm) in front: <!-- Decorator tigris ${decoTop} - ${decoBottom}-->
No problem with Firefox, the DOCTYPE is recognized there, despite the comment.

So, the html comment, <!-- Decorator ... should be commented out.

git-svn-id: http://svn.apache.org/repos/asf/portals/jetspeed-2/branches/JETSPEED-2.1.3@581798 13f79535-47bb-0310-9956-ffa450edef68