Synchronizing with trunk update of the MvnMojo:
If the .m2/settings.xml defines a custom </localRepository> location, executing jetspeed-mvn for a custom pom *and* a custom jetspeed-mvn-settings.xml,
the custom jetspeed-mvn-settings.xml also needs to be configured using that same </localRepository>, otherwise locally installled artifacts might not be resolved.
As this is quit a common case but easily forgotten, the currently defined localRepository path is now automatically passed onto the spawned maven execution for a custom pom as well. 

git-svn-id: http://svn.apache.org/repos/asf/portals/jetspeed-2/portal/branches/JS2-869@685736 13f79535-47bb-0310-9956-ffa450edef68