Moving back to static references of Spring beans within PortletEntityImpl to reduce the overhead of having to go through the (Spring) ComponentManager looking them up for each new instance. The major change now is that these static references are *not* directly set by those Spring beans onto the PortletEntityImpl but *injected* transparently through Spring.
This vastly reduces the coupling between these beans and much reliable lifecycle "management" of the references as they will be automatically nullified again when the Spring context is destroyed.
Clearing out the static references is very important to support reloading the Spring context (possibly with a new classloader) like happens with the (non-forked) test-cases and in the future hopefully will help making the Jetspeed portal itself reloadable.

This change for PortletEntityImpl is just one step: new functionality (like for the security refactoring) and other existing directly injected static references should be dealt with in the same way.  

git-svn-id: http://svn.apache.org/repos/asf/portals/jetspeed-2/portal/branches/security-refactoring@691681 13f79535-47bb-0310-9956-ffa450edef68