Renaming the USER_INFO_SUBSITE attribute constant name to JETSPEED_USER_SUBSITE_ATTRIBUTE *** and *** Changing its value to be properly namespaced from "subsite" to "org.apache.jetspeed.user.subsite"
Note: this is a data change which potentially might impact existing serializer seed data.
Because of that, we'll need to add specific migration support to the JetspeedSerializer to translate old style seed data for the old (USER_INFO) attribute "subsite"
to be mapped/stored under the name name.

To be able to keep track of such important still TODO migration support steps, I'll create a separate general JIRA issue
where we can record all the outstanding  migration support tasks.

git-svn-id: http://svn.apache.org/repos/asf/portals/jetspeed-2/portal/branches/security-refactoring@694070 13f79535-47bb-0310-9956-ffa450edef68