Complete implementation and proper handling of default supported JetspeedPrincipal association types. This includes isMemberOf, IsChildOf, IsA, IsPartOf, BelongsTo associations and their behaviors.
Also providing default configuration/support for the following associations:
- groupIsAGroup
- groupIsPartOfGroup
- roleIsARole
- roleIsPartOfRole
So, yes: *both* isA and IsPartOf can now be used at the same time for hierarical groups and/or roles!

And finally, the default configuration for the UserSubjectPrincipalsResolver actually evaluates all of the above properly:
- the children groups of isPartOf another group are collected (from side)
- the parent group which is the generalization of isA another group is collected (to side)
- the roles from groupIsMemberOfRole associations are collected
- the children roles of isPartOf another role are collected (from side)
- the parent role which is the generalization of isA another role is collected (to side)
- and all collected groups and roles are evaluated themselves as well (recursively)

One note: currently the j2-admin nor the JetspeedSerializer support defining isA or isPartOf associations for roles and groups.
The above has been verified and tested by manually creating these associations in the database.

git-svn-id: http://svn.apache.org/repos/asf/portals/jetspeed-2/portal/branches/security-refactoring@698741 13f79535-47bb-0310-9956-ffa450edef68