Reimplementation of the ServletContextProvider interface. With the new Pluto 2.0 container API this turned out to be somewhat more complex than anticipated but AFAIK everything now works as before which I have tested against Wicket 1.3.5.
However ... the current Struts Portlet Bridge (1.0.4) is *not* working anymore.
That is not so much caused by the ServletContextProvider implementation but Struts itself using the ServletContext.getRequestDispatcher instead of Request.getRequestDispatcher.
With the new Pluto 2.0, using a request dispatcher from the ServletContext and the fact that now ServletRequest != PortletRequest causes Struts to bypass the container completely
when doing *subsequent* includes. 
This really can only be fixed by upgrading the Struts Portlet Bridge itself. Until that happens, Struts Bridge based Portlets no longer can run on Jetspeed 2.2

git-svn-id: http://svn.apache.org/repos/asf/portals/jetspeed-2/portal/trunk@761258 13f79535-47bb-0310-9956-ffa450edef68