JS2-1091: backport of the feature:
Admin functionality to mark a portlet as out-of-service.
If administrator set a metadata, named "service.out", with value of "true" for a portlet definition, then the portlet will be displayed as "out-of-service".
The out-of-service message can be customized in the spring assembly configuration, "aggregation.xml", for the bean of "PortletRendererImpl" by adding the 'outOfServiceMessage' property like the following example:

        <property name="outOfServiceMessage">
            <value><![CDATA[
            <em>Sorry, this portlet is available now.</em>
            ]]></value>
        </property>

git-svn-id: http://svn.apache.org/repos/asf/portals/jetspeed-2/portal/branches/JETSPEED-2.1.4@893798 13f79535-47bb-0310-9956-ffa450edef68