Fixing NPE when there's no value assigned for a preference. This problem was pointed out by Ruben Carvalho. Thank you so much! Here's an explanation:
According to the portlet specification, it is possible for a portlet preference to be defined like this: 
<preference>
  <name>CategoryID</name>
</preference>

And, some codes like the following could NPE:

for (String s : pref.getValues()) { }

because PortletPreference#getValues() can return null instead of empty string array.
(For example, DatabasePreference returns null explicitly when it has empty value collection internally.)

Therefore, it is necessary to check if getValues() is null or not.

By the way, if you navigate portlet preference tab for a portlet, which has a non-value preference element, in Portlet Application Manager of j2-admin, then you can find that you cannot update the value for the preference.
However, you can still remove the preference and create a preference to have a value with that name again.
This seems a little strange, but anyway it is good enough for now.

git-svn-id: http://svn.apache.org/repos/asf/portals/jetspeed-2/portal/trunk@1002877 13f79535-47bb-0310-9956-ffa450edef68